pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool]
[tool.poetry]
authors = ["eggplants <w10776e8w@yahoo.co.jp>"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: Implementation :: CPython",
]
description = "Count LOC online"
keywords = ["loc", "web"]
name = "locloc"
packages = [{include = "locloc"}]
license = "MIT"
readme = "README.md"
repository = "https://github.com/eggplants/locloc"
version = "0.0.0"

[tool.poetry.dependencies]
python = ">=3.8,<4"
fastapi = "^0.103.1"
gitpython = "^3.1.35"
pytokei = "^0.2.0"
pydantic = "^2.3.0"

[tool.poetry.group.dev.dependencies]
mypy = ">=0.991,<1.6"
pre-commit = ">=2.20,<4.0"
taskipy = "^1.10.3"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
locloc-server = "locloc.main:main"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.ruff]
select = ["ALL"]
ignore = ["D", "ANN101"]
line-length = 120

[tool.ruff.mccabe]
max-complexity = 18

[tool.ruff.per-file-ignores]
"main.py" = [
  "T201", # `print` found
]
"tests/*test_*.py" = [
  "INP001",  # File tests/test_*.py is part of an implicit namespace package. Add an __init__.py.
  "S101", # Use of assert detected
]

[tool.mypy]
pretty = true
python_version = "3.9"
show_error_codes = true
strict = true

[tool.pytype]
python_version = '3.10'
jobs = '4'
report_errors = true
precise_return = true
protocols = true
strict_import = true

[tool.taskipy.tasks]
test = "pytest --cov=locloc --cov-report=term"
"test:ci" = "task test --cov-report=xml:cov.xml"
lint = "pre-commit run -a"
profile = "python -m cProfile"
